Correction animations

+++ common/heart_beat_thread.h
@@ -2,7 +2,7 @@
 #define HEART_BEAT_THREAD_H
 
 // Rythme cardiaque en 10ème de seconde
-extern volatile uint16_t heart_beat_speed;
+extern volatile int heart_beat_speed;
 
 void startHeartBeatThread(void);
